+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - terraform generator
 - DAP generator
+- Newrelic generator
 
 ## [0.1.0] - 2022-02-14